Splits off the Custom Rules region of EquivalencySpecs to its own file
