+++ FluentAssertions.Shared.Specs/Shared.Specs.projitems
@@ -26,6 +26,7 @@
     <Compile Include="$(MSBuildThisFileDirectory)EventAssertionSpecs.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ExceptionAssertionSpecs.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ExecutionTimeAssertionsSpecs.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)ExtensibilityRelatedEquivalencySpecs.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)FindAssembly.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)FluentDateTimeSpecs.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)FormatterSpecs.cs" />
